requirements/dev.txt
-r base.txt
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
ruff>=0.8,<1.0
django-debug-toolbar>=5.0,<6.0
